+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -113,7 +113,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-04-28T16:26:12.171Z")
+    updated: Optional[datetime] = parse_datetime("2026-04-29T12:22:32.001Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")