+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQAEKN31JGQH1T6MDVD2NDZS"
+        "01KQCK26YRWZ5MG2CQNTKAE2AW"
     )
     r"""The unique identifier of the data source"""