+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQAEKN32W9S61BGJ1FTJ6MCF"
+        "01KQCK26Z1Y7NXHVWG5VN4F3BX"
     )
     r"""The unique identifier of the data source"""